feature(server startup): create migration and seeding worker; migrate db -> add user table column isAdmin; seeding -> seed user Klaus, Chris and Christian with admin rights
