+++ app/server/migration/migrations.js
@@ -1,0 +1,5 @@
+module.exports = {
+    alterTableUsersAddIsAdmin: (dbConnection) => dbConnection.query(
+            `ALTER TABLE users ADD COLUMN IF NOT EXISTS "isAdmin" BOOLEAN;`
+        )
+}
